proxy: slim node agent from k8s and envoy dependencies
