+++ security/pkg/nodeagent/plugin/providers/google/stsclient/stsclient.go
@@ -24,10 +24,10 @@ import (
 	"net/http"
 	"time"
 
-	sec_model "istio.io/istio/pilot/pkg/security/model"
 	"istio.io/istio/pkg/bootstrap/platform"
 	"istio.io/istio/pkg/env"
 	"istio.io/istio/pkg/log"
+	sec_model "istio.io/istio/pkg/model"
 	"istio.io/istio/pkg/security"
 	"istio.io/istio/security/pkg/monitoring"
 )
